requirements.txt
ebird-api
pandas
pyperclip
